+++ solores/src/idl_format/anchor/mod.rs
@@ -1,15 +1,13 @@
-use heck::ToPascalCase;
-use proc_macro2::TokenStream;
-use quote::{format_ident, quote, ToTokens};
 use serde::Deserialize;
-
-use crate::gen_body_newtype_slice;
 
 use super::{IdlCodegenModule, IdlFormat};
 
 use self::{
-    accounts::NamedAccount, errors::ErrorEnumVariant, events::Event,
-    instructions::NamedInstruction, typedefs::NamedType,
+    accounts::{AccountsCodegenModule, NamedAccount},
+    errors::{ErrorEnumVariant, ErrorsCodegenModule},
+    events::{Event, EventsCodegenModule},
+    instructions::{IxCodegenModule, NamedInstruction},
+    typedefs::{NamedType, TypedefsCodegenModule},
 };
 
 mod accounts;
@@ -81,201 +79,4 @@ impl IdlFormat for AnchorIdl {
         }
         res
     }
-}
-
-struct AccountsCodegenModule<'a>(pub &'a [NamedAccount]);
-
-impl IdlCodegenModule for AccountsCodegenModule<'_> {
-    fn name(&self) -> &str {
-        "accounts"
-    }
-
-    fn gen_head(&self) -> TokenStream {
-        let mut res = quote! {
-            use borsh::{BorshDeserialize, BorshSerialize};
-        };
-        let mut has_pubkey = false;
-        let mut has_defined = false;
-        for a in self.0 {
-            if a.0.r#type.has_pubkey_field() && !has_pubkey {
-                has_pubkey = true;
-                res.extend(quote! {
-                    use solana_program::pubkey::Pubkey;
-                });
-            }
-            if a.0.r#type.has_defined_field() && !has_defined {
-                has_defined = true;
-                res.extend(quote! {
-                    use crate::*;
-                })
-            }
-            if has_defined && has_pubkey {
-                break;
-            }
-        }
-        res
-    }
-
-    gen_body_newtype_slice!();
-}
-
-struct TypedefsCodegenModule<'a>(pub &'a [NamedType]);
-
-impl IdlCodegenModule for TypedefsCodegenModule<'_> {
-    fn name(&self) -> &str {
-        "typedefs"
-    }
-
-    fn gen_head(&self) -> TokenStream {
-        let mut res = quote! {
-            use borsh::{BorshDeserialize, BorshSerialize};
-        };
-        for t in self.0 {
-            if t.r#type.has_pubkey_field() {
-                res.extend(quote! {
-                    use solana_program::pubkey::Pubkey;
-                });
-                break;
-            }
-        }
-        res
-    }
-
-    gen_body_newtype_slice!();
-}
-
-struct IxCodegenModule<'a>(pub &'a [NamedInstruction]);
-
-impl IdlCodegenModule for IxCodegenModule<'_> {
-    fn name(&self) -> &str {
-        "instructions"
-    }
-
-    fn gen_head(&self) -> TokenStream {
-        let mut res = quote! {
-            use borsh::{BorshDeserialize, BorshSerialize};
-            use solana_program::{
-                account_info::AccountInfo,
-                entrypoint::ProgramResult,
-                instruction::{AccountMeta, Instruction},
-                program::{invoke, invoke_signed},
-                pubkey::Pubkey,
-            };
-        };
-        for ix in self.0 {
-            if ix
-                .args
-                .iter()
-                .map(|a| a.r#type.is_or_has_defined())
-                .any(|b| b)
-            {
-                res.extend(quote! {
-                    use crate::*;
-                });
-                break;
-            }
-        }
-        res
-    }
-
-    gen_body_newtype_slice!();
-}
-
-struct ErrorsCodegenModule<'a> {
-    pub program_name: &'a str,
-    pub variants: &'a [ErrorEnumVariant],
-}
-
-impl IdlCodegenModule for ErrorsCodegenModule<'_> {
-    fn name(&self) -> &str {
-        "errors"
-    }
-
-    fn gen_head(&self) -> TokenStream {
-        quote! {
-            use solana_program::{
-                decode_error::DecodeError,
-                msg,
-                program_error::{PrintProgramError, ProgramError},
-            };
-            use thiserror::Error;
-        }
-    }
-
-    fn gen_body(&self) -> TokenStream {
-        let mut error_enum_variants = quote! {};
-        error_enum_variants.extend(self.variants.iter().map(|e| e.into_token_stream()));
-
-        let error_enum_ident_str = format!("{}Error", self.program_name.to_pascal_case());
-        let error_enum_ident = format_ident!("{}", &error_enum_ident_str);
-        quote! {
-            #[derive(Clone, Copy, Debug, Eq, Error, num_derive::FromPrimitive, PartialEq)]
-            pub enum #error_enum_ident {
-                #error_enum_variants
-            }
-
-            impl From<#error_enum_ident> for ProgramError {
-                fn from(e: #error_enum_ident) -> Self {
-                    ProgramError::Custom(e as u32)
-                }
-            }
-
-            impl<T> DecodeError<T> for #error_enum_ident {
-                fn type_of() -> &'static str {
-                    #error_enum_ident_str
-                }
-            }
-
-            impl PrintProgramError for #error_enum_ident {
-                fn print<E>(&self)
-                where
-                    E: 'static
-                        + std::error::Error
-                        + DecodeError<E>
-                        + PrintProgramError
-                        + num_traits::FromPrimitive,
-                {
-                    msg!(&self.to_string());
-                }
-            }
-        }
-    }
-}
-
-struct EventsCodegenModule<'a>(&'a [Event]);
-
-impl IdlCodegenModule for EventsCodegenModule<'_> {
-    fn name(&self) -> &str {
-        "events"
-    }
-
-    fn gen_head(&self) -> TokenStream {
-        let mut res = quote! {
-            use borsh::{BorshDeserialize, BorshSerialize};
-        };
-        let mut has_pubkey = false;
-        let mut has_defined = false;
-        for a in self.0 {
-            for field in &a.0.fields {
-                if field.r#type.is_or_has_pubkey() && !has_pubkey {
-                    has_pubkey = true;
-                    res.extend(quote! {
-                        use solana_program::pubkey::Pubkey;
-                    });
-                }
-                if field.r#type.is_or_has_defined() && !has_defined {
-                    has_defined = true;
-                    res.extend(quote! {
-                        use crate::*;
-                    })
-                }
-            }
-            if has_defined && has_pubkey {
-                break;
-            }
-        }
-        res
-    }
-
-    gen_body_newtype_slice!();
 }
